fixed compliation errors

+++ Plane.cs
@@ -1,7 +1,7 @@
 internal class Plane : SceneObject
 {
     public Vector Norm;
-    public double Offset;
+    public float Offset;
 
-    public Plane(Vector norm, double offset, Surface surface) : base(surface) { Norm = norm; Offset = offset; }
+    public Plane(Vector norm, float offset, Surface surface) : base(surface) { Norm = norm; Offset = offset; }
 }
